Refactor UI classes
Changes:
* Move CreateHexEditTab to MainWindow
* Rename View to IconAwareView
* Move createVisualization to MainWindow
* Rename showHexEditor to openHexEditor
* Move MainWindowWithDetachableDockWidgets to a separate file
* Add explaining comment to IconAwareView declaration
* Add copyright note to MainWindowWithDetachableDockWidgets
* Move ConnectionDialog to /dialogs
* Move CreateChunkDialog to /dialogs
* Move GoToAddressDialog to /dialogs
* Move SearchDialog to /dialogs
* Move OptionsDialog to /dialogs
* Move TabBarEventFilter to a new file
* Move ActivateDockEventFilter to a seperate file in /filters
* Move TabBarEventFilter to /filters
