+++ include/ui/filters/activatedockeventfilter.h
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2018 CodiLime
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+#pragma once
+
+#include <QEvent>
+#include <QObject>
+
+namespace veles {
+namespace ui {
+
+class ActivateDockEventFilter : public QObject {
+  Q_OBJECT
+
+ public:
+  explicit ActivateDockEventFilter(QObject* parent = nullptr);
+
+ protected:
+  bool eventFilter(QObject* watched, QEvent* event) override;
+};
+
+}  // namespace ui
+}  // namespace veles
