+++ include/ui/nodetreewidget.h
@@ -26,10 +26,10 @@
 
 #include "dbif/info.h"
 #include "dbif/types.h"
+#include "ui/dialogs/searchdialog.h"
 #include "ui/dockwidget.h"
 #include "ui/fileblobmodel.h"
 #include "ui/hexedit.h"
-#include "ui/searchdialog.h"
 #include "visualization/base.h"
 
 namespace veles {
