+++ include/ui/nodewidget.h
@@ -23,12 +23,12 @@
 #include <QToolButton>
 #include <QWidget>
 
+#include "ui/dialogs/searchdialog.h"
 #include "ui/dockwidget.h"
 #include "ui/fileblobmodel.h"
 #include "ui/hexedit.h"
 #include "ui/hexeditwidget.h"
 #include "ui/nodetreewidget.h"
-#include "ui/searchdialog.h"
 #include "util/sampling/uniform_sampler.h"
 #include "visualization/base.h"
 #include "visualization/minimap_panel.h"
