+++ src/ui/connectionmanager.cc
@@ -20,7 +20,7 @@
 #include <QFileDialog>
 #include <QTimer>
 
-#include "ui/connectiondialog.h"
+#include "ui/dialogs/connectiondialog.h"
 #include "ui/logwidget.h"
 #include "util/settings/connection_client.h"
 #include "util/settings/shortcuts.h"
