+++ src/ui/dialogs/connectiondialog.ui
