+++ src/ui/dialogs/searchdialog.ui
